+++ src/bignum/bignum_lshift.c
@@ -6,13 +6,13 @@
 /*   By: acazuc <acazuc@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/06 16:26:43 by acazuc            #+#    #+#             */
-/*   Updated: 2018/07/08 17:56:04 by acazuc           ###   ########.fr       */
+/*   Updated: 2018/08/04 21:04:12 by acazuc           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "bignum.h"
 
-int	bignum_lshift(t_bignum *r, t_bignum *a, uint64_t n)
+int	bignum_lshift(t_bignum *r, t_bignum *a, uint32_t n)
 {
 	int	sign;
 	int	ret;
