+++ src/bignum/bignum_lshift_op.c
@@ -6,7 +6,7 @@
 /*   By: acazuc <acazuc@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/08 17:54:57 by acazuc            #+#    #+#             */
-/*   Updated: 2018/07/09 21:05:38 by acazuc           ###   ########.fr       */
+/*   Updated: 2018/08/04 21:04:40 by acazuc           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ static int	do_init(t_bignum **tmp1, t_bignum **tmp2, t_bignum **tmp3)
 	return (1);
 }
 
-int	bignum_lshift_op(t_bignum *r, t_bignum *a, uint64_t n)
+int	bignum_lshift_op(t_bignum *r, t_bignum *a, uint32_t n)
 {
 	t_bignum	*tmp1;
 	t_bignum	*tmp2;
