+++ src/bignum/bignum_set_bit.c
@@ -6,16 +6,16 @@
 /*   By: acazuc <acazuc@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/09 19:34:54 by acazuc            #+#    #+#             */
-/*   Updated: 2018/07/09 20:52:11 by acazuc           ###   ########.fr       */
+/*   Updated: 2018/08/04 21:04:01 by acazuc           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "bignum.h"
 
-int	bignum_set_bit(t_bignum *bignum, uint64_t n)
+int	bignum_set_bit(t_bignum *bignum, uint32_t n)
 {
-	uint64_t	len;
-	uint64_t	old;
+	uint32_t	len;
+	uint32_t	old;
 
 	old = bignum->len;
 	len = 1 + n / 8 / sizeof(*bignum->data);
